BackEnd/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.1.0
langchain-ollama==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
